// packages/eslint-plugin-spellbookx/src/types.ts
export type SbxESLintConfig =
  | 'ignores'
  | 'javascript'
  | 'react'
  | 'json'
  | 'markdown'
  | 'cspell'
  | 'prettier'
  | 'recommended'
  | 'recommended-astro'
  | 'recommended-react'
  | 'recommended-no-spellcheck'
  | 'recommended-no-spellcheck-astro'
  | 'recommended-no-spellcheck-react';